requirements.txt
setuptools~=60.2.0
attrs~=21.4.0
multidict~=6.0.2
async-timeout~=4.0.2
aiohttp~=3.8.1
idna~=3.3
charset-normalizer~=2.1.0
py-cord~=2.0.0
yarl~=1.7.2
frozenlist~=1.3.0
aiosignal~=1.2.0
wheel~=0.37.1
python-dotenv